Error correction + dynamic bitrate support added
Only tested on the Jetson so far, but it works great. 

--nored  => Disable error correction. If you don't disable it, the bandwidth may be up to 2x higher than the target video bitrate.  I do not recommend removing, unless you're on a pristine connection.

--noqos => This will lower the bitrate of the video encoder if packet loss is detected. It won't lower it more than 5x (20% of target), but I find this works well to combat times where the network bandwidth is insufficient.  Error correction just doesn't work if it can't actually send enough data.

I lowered the default bitrate to 2500, which will be 5000-kbps with error correction on.  On a typical spotty WIFI connection, it might drop to 2000-kbps (4-mbps total bandwidth). 

I do NOT dynamically change the resolution, and 1500-kbps is about as low as 1080p will allow.  So, if on weak cellular or wifi, you might want to use 720p as the resolution. This should allow for a lower bitrate floor, allowing for use in bad environments.